Create base adapter with non-admin logic
